server/requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
pydantic
python.env
python-dotenv
flask-session
redis
flask-cors
redis